+++ card_live_dashboard/app.py
@@ -2,24 +2,33 @@ from typing import Union
 import dash
 import flask
 from pathlib import Path
-from os import path
+from os import getcwd
 
 from card_live_dashboard.service.CardLiveDataManager import CardLiveDataManager
 import card_live_dashboard.layouts as layouts
 import card_live_dashboard.callbacks as callbacks
 
-DEFAULT_DATA_DIR = Path(path.dirname(__file__), '..', 'data', 'card_live')
+DEFAULT_CARD_LIVE_HOME = Path(getcwd())
 
 
-def build_app(card_live_data_dir: Path = DEFAULT_DATA_DIR) -> dash.dash.Dash:
+def build_app(card_live_home: Path = DEFAULT_CARD_LIVE_HOME) -> dash.dash.Dash:
     """
     Builds the CARD:Live Dash application.
-    :param card_live_data_dir: The directory containing the CARD:Live data.
+    :param card_live_home: The home directory of CARD:Live.
     :return: The Dash application.
     """
+    if not card_live_home.exists():
+        raise Exception(f'Error: card live home [{card_live_home}] does not exist. Please specify the appropriate directory.')
+
+    card_live_data_dir = Path(card_live_home, 'data', 'card_live')
+    if not card_live_data_dir.exists():
+        raise Exception((f'Error: card live data directory [{card_live_data_dir}] does not exist. '
+                         'Please check for this directory or change card_live_home to something '
+                         f'other than [{card_live_home}]'))
+
     app = dash.Dash(name=__name__, external_stylesheets=layouts.external_stylesheets)
 
-    CardLiveDataManager.create_instance(card_live_data_dir)
+    CardLiveDataManager.create_instance(card_live_home)
 
     app.layout = layouts.default_layout()
     callbacks.build_callbacks(app)
@@ -27,13 +36,13 @@ def build_app(card_live_data_dir: Path = DEFAULT_DATA_DIR) -> dash.dash.Dash:
     return app
 
 
-def flask_app(data_dir: Union[str,Path] = DEFAULT_DATA_DIR) -> flask.Flask:
+def flask_app(card_live_home: Union[str,Path] = DEFAULT_CARD_LIVE_HOME) -> flask.Flask:
     """
     Builds and gets the Flask app object for deployment as a Flask app.
-    :param data_dir: The directory containing the data.
+    :param card_live_home: The home directory for CARD:Live.
     :return: The Flask app server object.
     """
-    if isinstance(data_dir, str):
-        data_dir = Path(data_dir)
+    if isinstance(card_live_home, str):
+        card_live_home = Path(card_live_home)
 
-    return build_app(data_dir).server
+    return build_app(card_live_home).server