+++ card_live_dashboard/callbacks/__init__.py
@@ -2,7 +2,6 @@ from typing import List, Set, Dict
 from datetime import datetime, timedelta
 from dash.dependencies import Input, Output, State
 import dash
-from dash.exceptions import PreventUpdate
 import re
 
 from card_live_dashboard.service.CardLiveDataManager import CardLiveDataManager
@@ -13,9 +12,14 @@ from card_live_dashboard.model import world
 DAY = timedelta(days=1)
 WEEK = timedelta(days=7)
 MONTH = timedelta(days=31)
-THREE_MONTHS = timedelta(days=365/4) # 3 months is defined as a quarter year
-SIX_MONTHS = timedelta(days=365/2) # 6 months is defined as half a year
+THREE_MONTHS = timedelta(days=365 / 4)  # 3 months is defined as a quarter year
+SIX_MONTHS = timedelta(days=365 / 2)  # 6 months is defined as half a year
 YEAR = timedelta(days=365)
+
+ORGANISM_COLUMN = {
+    'lmat': 'lmat_taxonomy',
+    'rgi_kmer': 'rgi_kmer_taxonomy',
+}
 
 
 def build_callbacks(app: dash.dash.Dash) -> None:
@@ -68,8 +72,7 @@ def build_callbacks(app: dash.dash.Dash) -> None:
          Output('time-period-items', 'options'),
          Output('date-picker-range', 'min_date_allowed'),
          Output('date-picker-range', 'max_date_allowed'),
-         Output('organism-lmat-select', 'options'),
-         Output('organism-rgi-kmer-select', 'options'),
+         Output('organism-select', 'options'),
          Output('selected-samples-count', 'children'),
          Output('drug-class-select', 'options'),
          Output('amr-gene-family-select', 'options'),
@@ -84,8 +87,8 @@ def build_callbacks(app: dash.dash.Dash) -> None:
          Input('amr-gene-family-select', 'value'),
          Input('resistance-mechanism-select', 'value'),
          Input('amr-gene-select', 'value'),
-         Input('organism-lmat-select', 'value'),
-         Input('organism-rgi-kmer-select', 'value'),
+         Input('organism-identification-method', 'value'),
+         Input('organism-select', 'value'),
          Input('time-period-items', 'value'),
          Input('date-picker-range', 'start_date'),
          Input('date-picker-range', 'end_date'),
@@ -98,9 +101,8 @@ def build_callbacks(app: dash.dash.Dash) -> None:
     )
     def update_all_figures(rgi_cutoff_select: str, drug_classes: List[str],
                            amr_gene_families: List[str], resistance_mechanisms: List[str],
-                           amr_genes: List[str], organism_lmat: str,
-                           organism_rgi_kmer: str, time_dropdown: str,
-                           start_date: str, end_date: str,
+                           amr_genes: List[str], organism_identification_method: str, organism: str,
+                           time_dropdown: str, start_date: str, end_date: str,
                            timeline_type_select: str, timeline_color_select: str,
                            totals_type_select: str, totals_color_select: str,
                            rgi_type_select: str, n_intervals):
@@ -139,17 +141,27 @@ def build_callbacks(app: dash.dash.Dash) -> None:
                                      amr_gene_families=amr_gene_families,
                                      resistance_mechanisms=resistance_mechanisms,
                                      amr_genes=amr_genes,
-                                     organism_lmat=organism_lmat,
-                                     organism_rgi_kmer=organism_rgi_kmer,
                                      custom_date=custom_date)
 
-        fig_settings = {
-            'timeline': {'type': timeline_type_select, 'color': timeline_color_select},
-            'totals': {'type': totals_type_select, 'color': totals_color_select},
-            'rgi': {'type': rgi_type_select}
-        }
+        # I have to extract the list of available organism options prior to filtering the data by the selected organism
+        # Otherwise once a user selects an organism there will be no other options available
+        organism_column = ORGANISM_COLUMN[organism_identification_method]
+        organism_options = build_options([organism],
+                                         time_subsets[time_dropdown].unique_column(organism_column))
 
-        main_pane_figures = build_main_pane(time_subsets[time_dropdown], fig_settings)
+        time_subsets = apply_organism_filter(time_subsets=time_subsets,
+                                             organism_identification_method=organism_identification_method,
+                                             organism=organism)
+
+        fig_settings = build_fig_settings(timeline_type_select=timeline_type_select,
+                                          timeline_color_select=timeline_color_select,
+                                          totals_type_select=totals_type_select,
+                                          totals_color_select=totals_color_select,
+                                          rgi_type_select=rgi_type_select,
+                                          organism_identification_method=organism_identification_method
+                                          )
+
+        main_pane_figures = build_main_pane(time_subsets[time_dropdown], organism_identification_method, fig_settings)
 
         # Set time dropdown text to include count of samples in particular time period
         # Should produce a list of dictionaries like [{'label': 'All (500)', 'value': 'all'}, ...]
@@ -163,16 +175,13 @@ def build_callbacks(app: dash.dash.Dash) -> None:
 
         samples_count_string = f'{time_subsets[time_dropdown].samples_count()}/{global_samples_count}'
 
-        organism_lmat_options = build_options([organism_lmat],
-                                              time_subsets[time_dropdown].unique_column('lmat_taxonomy'))
-        organism_rgi_kmer_options = build_options([organism_rgi_kmer],
-                                                  time_subsets[time_dropdown].unique_column('rgi_kmer_taxonomy'))
         drug_class_options = build_options(drug_classes,
                                            time_subsets[time_dropdown].rgi_parser.all_drugs())
         amr_gene_families_options = build_options(amr_gene_families,
                                                   time_subsets[time_dropdown].rgi_parser.all_amr_gene_family())
         resistance_mechanisms_options = build_options(resistance_mechanisms,
-                                                  time_subsets[time_dropdown].rgi_parser.all_resistance_mechanisms())
+                                                      time_subsets[
+                                                          time_dropdown].rgi_parser.all_resistance_mechanisms())
         amr_gene_options = build_options(amr_genes,
                                          time_subsets[time_dropdown].rgi_parser.all_amr_genes())
 
@@ -181,8 +190,7 @@ def build_callbacks(app: dash.dash.Dash) -> None:
                 time_dropdown_text,
                 min_date_allowed,
                 max_date_allowed,
-                organism_lmat_options,
-                organism_rgi_kmer_options,
+                organism_options,
                 samples_count_string,
                 drug_class_options,
                 amr_gene_families_options,
@@ -194,19 +202,36 @@ def build_callbacks(app: dash.dash.Dash) -> None:
                 main_pane_figures['rgi'])
 
 
+def build_fig_settings(timeline_type_select: str, timeline_color_select: str, totals_type_select: str,
+                       totals_color_select: str, rgi_type_select: str,
+                       organism_identification_method: str) -> Dict[str, Dict[str, str]]:
+    if timeline_color_select == 'organism':
+        timeline_color_select = f'{timeline_color_select}_{organism_identification_method}'
+    if totals_type_select == 'organism':
+        totals_type_select = f'{totals_type_select}_{organism_identification_method}'
+    if totals_color_select == 'organism':
+        totals_color_select = f'{totals_color_select}_{organism_identification_method}'
+
+    fig_settings = {
+        'timeline': {'type': timeline_type_select, 'color': timeline_color_select},
+        'totals': {'type': totals_type_select, 'color': totals_color_select},
+        'rgi': {'type': rgi_type_select}
+    }
+
+    return fig_settings
+
+
 def apply_filters(data: CardLiveData, rgi_cutoff_select: str,
                   drug_classes: List[str], amr_gene_families: List[str],
-                  resistance_mechanisms: List[str], amr_genes: List[str], organism_lmat: str,
-                  organism_rgi_kmer: str, custom_date: Dict[str, datetime]) -> Dict[str, CardLiveData]:
+                  resistance_mechanisms: List[str], amr_genes: List[str],
+                  custom_date: Dict[str, datetime]) -> Dict[str, CardLiveData]:
     time_now = datetime.now()
 
     data = data.select(table='rgi', by='cutoff', type='row', level=rgi_cutoff_select) \
         .select(table='rgi', by='drug', type='file', elements=drug_classes) \
         .select(table='rgi', by='amr_gene_family', type='file', elements=amr_gene_families) \
         .select(table='rgi', by='resistance_mechanism', type='file', elements=resistance_mechanisms) \
-        .select(table='rgi', by='amr_gene', type='file', elements=amr_genes) \
-        .select(table='main', by='lmat_taxonomy', taxonomy=organism_lmat) \
-        .select(table='main', by='rgi_kmer_taxonomy', taxonomy=organism_rgi_kmer)
+        .select(table='rgi', by='amr_gene', type='file', elements=amr_genes)
 
     time_subsets = {
         'all': data,
@@ -219,12 +244,22 @@ def apply_filters(data: CardLiveData, rgi_cutoff_select: str,
     }
 
     if custom_date is not None:
-        time_subsets['custom'] = data.select(table='main', by='time', start=custom_date['start'], end=custom_date['end'])
+        time_subsets['custom'] = data.select(table='main', by='time', start=custom_date['start'],
+                                             end=custom_date['end'])
     else:
         time_subsets['custom'] = data
 
     return time_subsets
 
+
+def apply_organism_filter(time_subsets: Dict[str, CardLiveData],
+                          organism_identification_method: str, organism: str) -> Dict[str, CardLiveData]:
+    time_subsets_filtered = {}
+    for key in time_subsets:
+        time_subsets_filtered[key] = time_subsets[key].select(
+            table='main', by=ORGANISM_COLUMN[organism_identification_method], taxonomy=organism)
+
+    return time_subsets_filtered
 
 def build_options(selected_options: List[str], all_available_options: Set[str]):
     if selected_options is None or len(selected_options) == 0:
@@ -243,7 +278,7 @@ def build_options(selected_options: List[str], all_available_options: Set[str]):
         all_available_options_set.union(selected_options_set))]
 
 
-def build_main_pane(data: CardLiveData, fig_settings: Dict[str, Dict[str, str]]):
+def build_main_pane(data: CardLiveData, organism_identification_method: str, fig_settings: Dict[str, Dict[str, str]]):
     fig_map = figures.choropleth_drug(data, world)
     fig_histogram_rate = figures.build_time_histogram(data, fig_type=fig_settings['timeline']['type'],
                                                       color_by=fig_settings['timeline']['color'])