+++ card_live_dashboard/model/data_modifiers/AddTaxonomyModifier.py
@@ -1,4 +1,6 @@
 import logging
+from ete3 import NCBITaxa
+from pathlib import Path
 
 from card_live_dashboard.model.data_modifiers.CardLiveDataModifier import CardLiveDataModifier
 from card_live_dashboard.model.CardLiveData import CardLiveData
@@ -10,15 +12,18 @@ logger = logging.getLogger(__name__)
 
 class AddTaxonomyModifier(CardLiveDataModifier):
 
-    def __init__(self):
+    def __init__(self, ncbi_taxa_db: NCBITaxa):
         """
         Builds a new modifier which will add in taxonomy categories.
+
+        :param ncbi_taxa_db: The ete3 NCBI taxa database object.
         """
         super().__init__()
+        self._ncbi_taxa = ncbi_taxa_db
 
     def modify(self, data: CardLiveData) -> CardLiveData:
         logger.debug(f'Main df before {data.main_df}')
-        taxonomy_parser = TaxonomicParser(df_rgi_kmer=data.rgi_kmer_df, df_lmat=data.lmat_df)
+        taxonomy_parser = TaxonomicParser(self._ncbi_taxa, df_rgi_kmer=data.rgi_kmer_df, df_lmat=data.lmat_df)
         matches_df = taxonomy_parser.create_file_matches().rename(
             columns={'lmat.taxonomy_label': 'lmat_taxonomy',
                      'rgi_kmer.taxonomy_label': 'rgi_kmer_taxonomy'