+++ card_live_dashboard/test/unit/service/test_TaxonomicParser.py
@@ -1,13 +1,46 @@
+from typing import List, Dict, Optional
 import pandas as pd
+from ete3 import NCBITaxa
 
 from card_live_dashboard.service.TaxonomicParser import TaxonomicParser
 
 
+# There's likely a way to use mock objects here instead of just creating a new class with the same methods
+# as NCBITaxa, but I couldn't figure out how to do it. That is applying @unittest.mock.patch('ete3.NCBITaxa') wouldn't
+# actually mock ete3.NCBITaxa.
+# I also know there's a warning here that I'm not calling the superclass constructor, but I can't call it since it will
+# attempt to install the entire NCBI Taxonomy database, which I don't want in unit tests.
+class NCBITaxaMock(NCBITaxa):
+    LINEAGES = {
+        28901: 'Salmonella enterica',
+        543: 'Enterobacteriaceae',
+        1639: 'Listeria monocytogenes',
+    }
+
+    def __init__(self):
+        pass
+
+    def get_lineage(self, lineage: int) -> Optional[List[int]]:
+        return [lineage]
+
+    def get_rank(self, lineages: List[int]) -> Dict[int, str]:
+        ranks = {}
+        for lineage in lineages:
+            ranks[lineage] = 'species'
+        return ranks
+
+    def get_taxid_translator(self, taxids: List[int], try_synonyms: bool = True) -> Dict[int, str]:
+        names = {}
+        for id in taxids:
+            names[id] = self.LINEAGES[id]
+        return names
+
+
 def test_taxonomic_parser_simple():
-    lmat_df = pd.DataFrame(columns=['filename', 'lmat.taxonomy_label', 'lmat.count'],
+    lmat_df = pd.DataFrame(columns=['filename', 'lmat.taxonomy_label', 'lmat.count', 'lmat.ncbi_taxon_id'],
                            data=[
-                               ['file1', 'Salmonella enterica', '10'],
-                               ['file1', 'Salmonella enterica', '10'],
+                               ['file1', 'Salmonella enterica', '10', 28901],
+                               ['file1', 'Salmonella enterica', '10', 28901],
                            ]).set_index('filename')
 
     rgi_kmer_df = pd.DataFrame(columns=['filename', 'rgi_kmer.CARD*kmer Prediction'],
@@ -16,7 +49,7 @@ def test_taxonomic_parser_simple():
                                    ['file1', 'Salmonella enterica (chromosome)'],
                                ]).set_index('filename')
 
-    tax = TaxonomicParser(df_lmat=lmat_df, df_rgi_kmer=rgi_kmer_df)
+    tax = TaxonomicParser(ncbi_taxa=NCBITaxaMock(), df_lmat=lmat_df, df_rgi_kmer=rgi_kmer_df)
     file_matches = tax.create_file_matches()
 
     assert 1 == len(file_matches)
@@ -26,10 +59,10 @@ def test_taxonomic_parser_simple():
 
 
 def test_taxonomic_parser_no_match():
-    lmat_df = pd.DataFrame(columns=['filename', 'lmat.taxonomy_label', 'lmat.count'],
+    lmat_df = pd.DataFrame(columns=['filename', 'lmat.taxonomy_label', 'lmat.count', 'lmat.ncbi_taxon_id'],
                            data=[
-                               ['file1', 'Salmonella enterica', '10'],
-                               ['file1', 'Enterobacteriaceae', '20'],
+                               ['file1', 'Salmonella enterica', '10', 28901],
+                               ['file1', 'Enterobacteriaceae', '20', 543],
                            ]).set_index('filename')
 
     rgi_kmer_df = pd.DataFrame(columns=['filename', 'rgi_kmer.CARD*kmer Prediction'],
@@ -38,7 +71,7 @@ def test_taxonomic_parser_no_match():
                                    ['file1', 'Salmonella enterica (plasmid)'],
                                ]).set_index('filename')
 
-    tax = TaxonomicParser(df_lmat=lmat_df, df_rgi_kmer=rgi_kmer_df)
+    tax = TaxonomicParser(ncbi_taxa=NCBITaxaMock(), df_lmat=lmat_df, df_rgi_kmer=rgi_kmer_df)
     file_matches = tax.create_file_matches()
 
     assert 1 == len(file_matches)
@@ -48,10 +81,10 @@ def test_taxonomic_parser_no_match():
 
 
 def test_taxonomic_parser_multiple_match():
-    lmat_df = pd.DataFrame(columns=['filename', 'lmat.taxonomy_label', 'lmat.count'],
+    lmat_df = pd.DataFrame(columns=['filename', 'lmat.taxonomy_label', 'lmat.count', 'lmat.ncbi_taxon_id'],
                            data=[
-                               ['file1', 'Salmonella enterica', '20'],
-                               ['file1', 'Enterobacteriaceae', '10'],
+                               ['file1', 'Salmonella enterica', '20', 28901],
+                               ['file1', 'Enterobacteriaceae', '10', 543],
                            ]).set_index('filename')
 
     rgi_kmer_df = pd.DataFrame(columns=['filename', 'rgi_kmer.CARD*kmer Prediction'],
@@ -60,7 +93,7 @@ def test_taxonomic_parser_multiple_match():
                                    ['file1', 'Salmonella enterica (plasmid)'],
                                ]).set_index('filename')
 
-    tax = TaxonomicParser(df_lmat=lmat_df, df_rgi_kmer=rgi_kmer_df)
+    tax = TaxonomicParser(ncbi_taxa=NCBITaxaMock(), df_lmat=lmat_df, df_rgi_kmer=rgi_kmer_df)
     file_matches = tax.create_file_matches()
 
     assert 1 == len(file_matches)
@@ -70,12 +103,12 @@ def test_taxonomic_parser_multiple_match():
 
 
 def test_taxonomic_parser_multiple_files():
-    lmat_df = pd.DataFrame(columns=['filename', 'lmat.taxonomy_label', 'lmat.count'],
+    lmat_df = pd.DataFrame(columns=['filename', 'lmat.taxonomy_label', 'lmat.count', 'lmat.ncbi_taxon_id'],
                            data=[
-                               ['file1', 'Salmonella enterica', '20'],
-                               ['file1', 'Enterobacteriaceae', '10'],
-                               ['file2', 'Enterobacteriaceae', '20'],
-                               ['file2', 'Salmonella enterica', '10'],
+                               ['file1', 'Salmonella enterica', '20', 28901],
+                               ['file1', 'Enterobacteriaceae', '10', 543],
+                               ['file2', 'Enterobacteriaceae', '20', 543],
+                               ['file2', 'Salmonella enterica', '10', 28901],
                            ]).set_index('filename')
 
     rgi_kmer_df = pd.DataFrame(columns=['filename', 'rgi_kmer.CARD*kmer Prediction'],
@@ -86,7 +119,7 @@ def test_taxonomic_parser_multiple_files():
                                    ['file2', 'Enterobacteriaceae (plasmid)'],
                                ]).set_index('filename')
 
-    tax = TaxonomicParser(df_lmat=lmat_df, df_rgi_kmer=rgi_kmer_df)
+    tax = TaxonomicParser(ncbi_taxa=NCBITaxaMock(), df_lmat=lmat_df, df_rgi_kmer=rgi_kmer_df)
     file_matches = tax.create_file_matches()
 
     assert 2 == len(file_matches)
@@ -96,11 +129,11 @@ def test_taxonomic_parser_multiple_files():
 
 
 def test_taxonomic_parser_tie_case1():
-    lmat_df = pd.DataFrame(columns=['filename', 'lmat.taxonomy_label', 'lmat.count'],
+    lmat_df = pd.DataFrame(columns=['filename', 'lmat.taxonomy_label', 'lmat.count', 'lmat.ncbi_taxon_id'],
                            data=[
-                               ['file1', 'Salmonella enterica', '20'],
-                               ['file1', 'Enterobacteriaceae', '20'],
-                               ['file1', 'Listeria monocytogenes', '10'],
+                               ['file1', 'Salmonella enterica', '20', 28901],
+                               ['file1', 'Enterobacteriaceae', '20', 543],
+                               ['file1', 'Listeria monocytogenes', '10', 1639],
                            ]).set_index('filename')
 
     rgi_kmer_df = pd.DataFrame(columns=['filename', 'rgi_kmer.CARD*kmer Prediction'],
@@ -112,7 +145,7 @@ def test_taxonomic_parser_tie_case1():
                                    ['file1', 'Listeria monocytogenes (chromosome)'],
                                ]).set_index('filename')
 
-    tax = TaxonomicParser(df_lmat=lmat_df, df_rgi_kmer=rgi_kmer_df)
+    tax = TaxonomicParser(ncbi_taxa=NCBITaxaMock(), df_lmat=lmat_df, df_rgi_kmer=rgi_kmer_df)
     file_matches = tax.create_file_matches()
 
     assert 1 == len(file_matches)
@@ -122,11 +155,11 @@ def test_taxonomic_parser_tie_case1():
 
 
 def test_taxonomic_parser_tie_case2():
-    lmat_df = pd.DataFrame(columns=['filename', 'lmat.taxonomy_label', 'lmat.count'],
+    lmat_df = pd.DataFrame(columns=['filename', 'lmat.taxonomy_label', 'lmat.count', 'lmat.ncbi_taxon_id'],
                            data=[
-                               ['file1', 'Enterobacteriaceae', '20'],
-                               ['file1', 'Listeria monocytogenes', '10'],
-                               ['file1', 'Salmonella enterica', '20'],
+                               ['file1', 'Enterobacteriaceae', '20', 543],
+                               ['file1', 'Listeria monocytogenes', '10', 1639],
+                               ['file1', 'Salmonella enterica', '20', 28901],
                            ]).set_index('filename')
 
     rgi_kmer_df = pd.DataFrame(columns=['filename', 'rgi_kmer.CARD*kmer Prediction'],
@@ -138,7 +171,7 @@ def test_taxonomic_parser_tie_case2():
                                    ['file1', 'Enterobacteriaceae (chromosome)'],
                                ]).set_index('filename')
 
-    tax = TaxonomicParser(df_lmat=lmat_df, df_rgi_kmer=rgi_kmer_df)
+    tax = TaxonomicParser(ncbi_taxa=NCBITaxaMock(), df_lmat=lmat_df, df_rgi_kmer=rgi_kmer_df)
     file_matches = tax.create_file_matches()
 
     assert 1 == len(file_matches)